refactor: rename jotaiInit and update imports for consistency
- Renamed jotaiInit to jotaiInitFromUi for clarity.
- Updated import paths in uiJsBridge and jotaiBgSync to reflect the new file structure.
- Introduced jotaiInitFromUi.ts to encapsulate the updated functionality.
