+++ packages/kit-bg/src/states/jotai/jotaiBgSync.ts
@@ -3,7 +3,7 @@ import type { IGlobalStatesSyncBroadcastParams } from '@onekeyhq/shared/src/back
 import { GLOBAL_STATES_SYNC_BROADCAST_METHOD_NAME } from '@onekeyhq/shared/src/background/backgroundUtils';
 import platformEnv from '@onekeyhq/shared/src/platformEnv';
 
-import { jotaiInitFromUi } from './jotaiInit';
+import { jotaiInitFromUi } from './jotaiInitFromUi';
 
 import type { EAtomNames } from './atomNames';
 import type BackgroundApiProxy from '../../apis/BackgroundApiProxy';
