+++ packages/kit-bg/src/states/jotai/jotaiInit.ts
@@ -1,10 +1,8 @@
 import { isNil, isPlainObject } from 'lodash';
 
 import appGlobals from '@onekeyhq/shared/src/appGlobals';
-import type { IGlobalStatesSyncBroadcastParams } from '@onekeyhq/shared/src/background/backgroundUtils';
 
 import localDb from '../../dbs/local/localDb';
-import { ELocalDBStoreNames } from '../../dbs/local/localDBStoreNames';
 
 import { EAtomNames } from './atomNames';
 import {
@@ -15,51 +13,7 @@ import {
 import { JotaiCrossAtom } from './utils/JotaiCrossAtom';
 import { jotaiDefaultStore } from './utils/jotaiDefaultStore';
 
-import type { IJotaiAtomSetWithoutProxy, IJotaiWritableAtomPro } from './types';
-
-export async function jotaiUpdateFromUiByBgBroadcast(
-  params: IGlobalStatesSyncBroadcastParams,
-) {
-  const allAtoms = await import('./atoms');
-  // @ts-ignore
-  // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
-  const atomInfo = allAtoms[params.name] as JotaiCrossAtom<any>;
-  // eslint-disable-next-line @typescript-eslint/no-unsafe-call
-  const atomObj = atomInfo.atom() as unknown as IJotaiWritableAtomPro<
-    any,
-    any,
-    any
-  >;
-  await jotaiDefaultStore.set(atomObj, params);
-}
-
-export async function jotaiInitFromUi({
-  states,
-}: {
-  states: Record<EAtomNames, any>;
-}) {
-  const allAtoms = await import('./atoms');
-  await Promise.all(
-    Object.entries(states).map(async ([key, value]) => {
-      // @ts-ignore
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
-      const atomInfo = allAtoms[key] as JotaiCrossAtom<any>;
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-call
-      const atomObj = atomInfo.atom() as unknown as IJotaiWritableAtomPro<
-        any,
-        any,
-        any
-      >;
-      const data: IJotaiAtomSetWithoutProxy = {
-        $$isForceSetAtomWithoutProxy: true,
-        name: key,
-        payload: value,
-      };
-      await jotaiDefaultStore.set(atomObj, data);
-    }),
-  );
-  globalJotaiStorageReadyHandler.resolveReady(true);
-}
+import type { IJotaiWritableAtomPro } from './types';
 
 function checkAtomNameMatched(key: string, value: string) {
   if (key !== value) {
